+++ karaage/common/forms.py
@@ -20,10 +20,11 @@
 
 import six
 from django import forms
+from django.conf import settings
+import re
 
 from karaage.common.models import LogEntry, COMMENT
 from .passwords import assert_strong_password
-
 
 def validate_password(username, password1, password2=None, old_password=None):
     # password1 is mandatory, it must be given in order to proceed.
@@ -46,6 +47,32 @@ def validate_password(username, password1, password2=None, old_password=None):
     # If password1 is ok, return it.
     return password1
 
+def clean_email(email):
+    email_match_type = "exclude"
+    email_match_list = []
+    if hasattr(settings, 'EMAIL_MATCH_TYPE'):
+        email_match_type = settings.EMAIL_MATCH_TYPE
+    if hasattr(settings, 'EMAIL_MATCH_LIST'):
+        email_match_list = settings.EMAIL_MATCH_LIST
+
+    found = False
+    for string in email_match_list:
+        m = re.search(string, email, re.IGNORECASE)
+        if m is not None:
+            found = True
+            break
+
+    message = "This email address cannot be used."
+    if hasattr(settings, 'EMAIL_MATCH_MSG'):
+        message = settings.EMAIL_MATCH_MSG
+    if email_match_type == "include":
+        if not found:
+            raise forms.ValidationError(message)
+    elif email_match_type == "exclude":
+        if found:
+            raise forms.ValidationError(message)
+    else:
+        raise forms.ValidationError("Oops. Nothing is valid. Sorry.")
 
 class CommentForm(forms.ModelForm):
     """ Comment form. """