+++ karaage/people/forms.py
@@ -32,7 +32,7 @@ from karaage.projects.models import Project
 from karaage.projects.utils import add_user_to_project
 from karaage.common.constants import COUNTRIES
 from karaage.common import get_current_person
-from karaage.common.forms import validate_password
+from karaage.common.forms import validate_password, clean_email
 
 
 class PersonForm(forms.ModelForm):
@@ -91,6 +91,20 @@ class PersonForm(forms.ModelForm):
             'supervisor', 'department', 'telephone', 'mobile', 'fax',
             'address', 'country'
         ]
+
+    def clean_email(self):
+        email = self.cleaned_data['email']
+        users = Person.objects.filter(email__exact=email)
+        if self.instance:
+            users = users.exclude(pk=self.instance.pk)
+        if users.count() > 0:
+            raise forms.ValidationError(
+                six.u(
+                    'An account with this email already exists. '
+                    'Please email %s')
+                % settings.ACCOUNTS_EMAIL)
+        clean_email(email)
+        return email
 
 
 class AdminPersonForm(PersonForm):